+++ .coveralls.yml
@@ -1,0 +1,1 @@
+service_name: travis-ci
