+++ .travis.yml
@@ -13,9 +13,7 @@ install:
 - npm install
 
 script:
-- ng build
-- ng test
+- ng test --watch false
 
-branches:
-  only:
-    - master
+after_script:
+- cat coverage/lcov.info | node_modules/coveralls/bin/coveralls.js
