+++ src/app/exemples/admin/admin.ts
@@ -1,7 +1,7 @@
 import {Component} from '@angular/core';
 import {bootstrap} from '@angular/platform-browser-dynamic';
-import {HmAdminComponent} from '../../hm-admin/hm-admin.component';
 import configuration from './configuration';
+import {HmAdminComponent} from '../../hm-admin/hm-admin.component';
 
 @Component({
   selector: 'app',
